account 앱 테스트 모듈화
테스팅 시 conftest를 사용하여 소스를 줄이고 앱 단위로 테스팅 구조화
